requirements.txt
setuptools==70.1.1
wheel>=0.44
build>=1.2.2
packaging>=24.1

langchain>=0.3.1
langchain-community>=0.3.1
langchain-text-splitters>=0.3.0
langchain-openai>=0.2.0
langchain-ollama>=0.2.0
langchain-huggingface>=0.1.1

faiss-cpu>=1.8.0
chromadb==1.1.1
langchain-chroma==0.2.6
# (Let chromadb pull PyPika itself; we’ll control build isolation via CLI/env)

pypdf>=5.1.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
python-dotenv>=1.0.1
requests>=2.32.3
httpx>=0.27.2
sentence-transformers>=3.0.1
tqdm>=4.66.4

bs4
#langchain groq
langchain-groq
langchain_core

